requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
colorama>=0.4.4
rich>=12.0.0

# Data processing
pandas>=1.3.0
jinja2>=3.0.0

# Async & Threading
asyncio>=3.4.3
concurrent-futures>=3.1.1

# Validation
validators>=0.18.0
ipaddress>=1.0.23

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0

# Linting & Formatting
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0